@@ -1,6 +1,8 @@
-require_relative 'boot'
+# frozen_string_literal: true
 
-#require 'rails/all'
+require_relative "boot"
+
+# require 'rails/all'
 
 require "decidim/rails"
 
@@ -18,7 +20,7 @@ Bundler.require(*Rails.groups)
 module DecidimApp
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.2
+    config.load_defaults 6.1
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
