@@ -1,16 +1,6 @@
-ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../Gemfile', __dir__)
+# frozen_string_literal: true
 
-require 'bundler/setup' # Set up gems listed in the Gemfile.
-require "bootsnap"
+ENV["BUNDLE_GEMFILE"] ||= File.expand_path("../Gemfile", __dir__)
 
-env = ENV["RAILS_ENV"] || "development"
-
-Bootsnap.setup(
-  cache_dir: File.expand_path(File.join("..", "tmp", "cache"), __dir__),
-  development_mode: env == "development",
-  load_path_cache: true,
-  autoload_paths_cache: true,
-  disable_trace: false,
-  compile_cache_iseq: !ENV["SIMPLECOV"],
-  compile_cache_yaml: true
-)
+require "bundler/setup" # Set up gems listed in the Gemfile.
+require "bootsnap/setup" # Speed up boot time by caching expensive operations.
