@@ -1,12 +1,14 @@
+# frozen_string_literal: true
+
+require "sidekiq/web"
+require "sidekiq/cron/web"
+
 Rails.application.routes.draw do
-  if Rails.env.development?
-    mount LetterOpenerWeb::Engine, at: "/letter_opener"
+  mount LetterOpenerWeb::Engine, at: "/letter_opener" if Rails.env.development?
+  authenticate :user, ->(u) { u.admin? } do
+    mount Sidekiq::Web => "/sidekiq"
   end
 
-  mount Decidim::Core::Engine => '/'
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-
-  # Añadido para el file autorizatión handler
-  mount Decidim::FileAuthorizationHandler::AdminEngine => '/admin'
-
+  mount Decidim::Core::Engine => "/"
+  mount Decidim::FileAuthorizationHandler::AdminEngine => "/admin"
 end
