@@ -1,5 +1,5 @@
-# frozen_string_literal: true
 # This migration comes from decidim_consultations (originally 20180115170933)
+# frozen_string_literal: true
 
 class AddSlugToDecidimConsultationsQuestions < ActiveRecord::Migration[5.1]
   class Question < ApplicationRecord
@@ -11,7 +11,7 @@ class AddSlugToDecidimConsultationsQuestions < ActiveRecord::Migration[5.1]
                :decidim_organization_id,
                :integer,
                index: {
-                 name: "index_decidim_questions_on_decidim_organization_id"
+                 name: 'index_decidim_questions_on_decidim_organization_id'
                }
 
     add_column :decidim_consultations_questions, :slug, :string
